+++ src/main/java/SopraAJC/NotreProjet/restcontroller/PartieRestController.java
@@ -36,7 +36,7 @@ public class PartieRestController {
 
     @GetMapping("/{id}")
     @JsonView(JsonViews.PartieWithSession.class)
-    public Partie partieById(@PathVariable("id") Integer id){
+    public Partie partieById(@PathVariable("id") Long id){
         partieService.partieExistsInDB(id);
         return partieRepository.findById(id).get();
     }
